+++ danny-store/src/main/java/danny/store/dannystore/domain/dto/OrderDtoForAdmin.java
@@ -7,7 +7,7 @@ public class OrderDtoForAdmin {
     private Long id;
     private String src;
     private String listProducts;
-    private Float totalAmount;
+    private Long totalAmount;
     private String username;
     private String nameCustomer;
     private String phoneNumber;
