+++ danny-store/src/main/java/danny/store/dannystore/domain/entity/Order.java
@@ -18,7 +18,7 @@ public class Order {
     private Long id;
     private Long saleStaffId;
     private Long customerId;
-    private Float totalAmount;
+    private Long totalAmount;
     private Long statusId;
     private String deliveryAddress;
     private String phone;
