+++ danny-store/src/main/java/danny/store/dannystore/domain/entity/OrderItem.java
@@ -20,7 +20,7 @@ public class OrderItem {
     private Long id;
     private Long orderId;
     private Long productId;
-    private Float price;
+    private Long price;
     private Long quantity;
     private Date createdAt;
     private Date modifiedAt;
