+++ danny-store/src/main/java/danny/store/dannystore/domain/entity/Promotion.java
@@ -16,7 +16,7 @@ public class Promotion {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
-    private Float percentValue;
+    private Long percentValue;
     private String namePromotion;
     private String descriptionPromotion;
     private Date startDayPromotion;
